refactor: extract syntax rewriters
